backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
huggingface_hub
python-dotenv
